requirements.txt
aniso8601==7.0.0
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
flask==3.0.0
Flask-GraphQL==2.0.1
Flask-JWT-Extended==4.5.3
graphene==2.1.9
graphene-mongo==0.1.19
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
idna==3.4
importlib-metadata==6.8.0
iso8601==2.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mongoengine==0.27.0
pandas==2.1.2
promise==2.3
PyJWT==2.8.0
pymongo==4.6.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
Rx==1.6.3
singledispatch==4.1.0
six==1.16.0
typing==3.7.4.3
tzdata==2023.3
urllib3==2.0.7
werkzeug==3.0.1
zipp==3.17.0
